refactor(web): split personal assistant mega state components
Extract running-state and empty-state UI into focused subfolder modules
so chat/confirmation/composer and landing journey/examples can change
without editing multi-thousand-line files. Import paths unchanged.
