+++ apps/web/src/app/(app)/personal-assistant/components/running-state.tsx
@@ -1,100 +1,28 @@
 "use client";
 
-import {
-  AlertCircle,
-  ArrowDown,
-  ArrowUpRight,
-  Building2,
-  Check,
-  ChevronRight,
-  Coins,
-  Copy,
-  Loader2,
-  Plus,
-  Settings as SettingsIcon,
-  Wrench,
-  X,
-  Zap,
-} from "lucide-react";
-import Link from "next/link";
+import { ArrowDown, Loader2 } from "lucide-react";
 import { useSearchParams } from "next/navigation";
-import { useFormatter, useTranslations } from "next-intl";
-import React, {
-  createContext,
+import { useTranslations } from "next-intl";
+import {
   type FormEvent,
   useCallback,
-  useContext,
   useEffect,
   useRef,
   useState,
 } from "react";
 import { toast } from "sonner";
 
-import {
-  applyConfirmationOrgProposalUpdate,
-  buildConfirmationApproveOverrideIfChanged,
-  buildCurrentConfirmationApproveOrganizationOverride,
-  CONFIRMATION_PERSONAL_SCOPE_VALUE,
-  isConfirmationOrgAwareTool,
-  mergeConfirmationOrgPickerOptions,
-  resolveConfirmationOrgPickerValue,
-} from "@/app/personal-assistant/components/confirmation-org-picker";
-import RotatingMessages from "@/app/personal-assistant/components/rotating-messages";
 import SettingsPanel from "@/app/personal-assistant/components/settings-panel";
-import { AuroraOrb } from "@/components/aurora-orb";
-import { ArrowUpIcon, StopIcon } from "@/components/chat/icons";
 import {
-  PromptInput,
-  PromptInputSubmit,
-  PromptInputTextarea,
-  PromptInputToolbar,
-  PromptInputTools,
-} from "@/components/chat/prompt-input";
-import Markdown from "@/components/markdown";
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
-import { Button } from "@/components/ui/button";
-import {
-  FileUpload,
-  FileUploadDropzone,
-  FileUploadItem,
-  FileUploadItemDelete,
-  FileUploadItemMetadata,
-  FileUploadItemPreview,
-  FileUploadList,
-  FileUploadTrigger,
-} from "@/components/ui/file-upload";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipTrigger,
-} from "@/components/ui/tooltip";
-import {
-  approveHermesConfirmationAction,
   listHermesMessagesAction,
   markHermesInboxSeenAction,
-  rejectHermesConfirmationAction,
 } from "@/lib/actions/hermes";
-import type { OrbExpression } from "@/lib/aurora-orb";
 import {
   HERMES_CONFIRMATION_CARD_KIND,
   parseConfirmationCardMessage,
 } from "@/lib/hermes/confirmation-card-message";
-import {
-  type HermesUiMessage,
-  mergeHermesMessageLists,
-} from "@/lib/hermes/merge-persisted-messages";
-import {
-  DEFAULT_ORB_MOTION,
-  type OrbMotion,
-  personalityToOrbMotion,
-} from "@/lib/hermes/personality-orb";
+import { mergeHermesMessageLists } from "@/lib/hermes/merge-persisted-messages";
+import { personalityToOrbMotion } from "@/lib/hermes/personality-orb";
 import {
   deltaContentFrom,
   type HermesStatusEvent,
@@ -108,19 +36,44 @@ import type {
   HermesPersistedMessage,
 } from "@/lib/hermes/types";
 import { orderedMessageList } from "@/lib/intl/ordered-message-list";
-import { cn } from "@/lib/utils";
 import AutonomyPanel from "./autonomy-panel";
+import {
+  AssistantMotionContext,
+  AssistantSeedContext,
+} from "./running-state/assistant-context";
+import { Composer } from "./running-state/composer";
+import { ConfirmationCard } from "./running-state/confirmation-card";
+import { buildMockPendingConfirmations } from "./running-state/confirmation-mock";
+import {
+  HERMES_STREAMING_ENABLED,
+  POLL_INTERVAL_MS,
+  REASONING_MIN_MS,
+} from "./running-state/constants";
+import { AutonomyChip, IntegrationsChip } from "./running-state/header-chips";
+import {
+  clientMimeForHermesUpload,
+  durationKey,
+  fileToDataUrl,
+  hasSameMessageIds,
+  persistedToMessage,
+  persistedToMessages,
+} from "./running-state/message-helpers";
+import { MessageRow } from "./running-state/message-row";
+import {
+  AssistantTyping,
+  ProgressChips,
+  ReasoningLine,
+} from "./running-state/progress-ui";
+import type {
+  ChatApiResponse,
+  Message,
+  ProgressStep,
+  ResolvedConfirmationEntry,
+  TimelineEntry,
+} from "./running-state/types";
+import { WelcomeBlock } from "./running-state/welcome-block";
 
-/** The committed orb seed — the assistant's avatar across the chat. */
-const AssistantSeedContext = createContext<string>("personal-assistant");
-
-/**
- * The orb's personality-driven motion (liveliness + resting eyes), published to
- * the chat avatars the same way the seed is. Defaults to calm/neutral.
- */
-const AssistantMotionContext = createContext<OrbMotion>(DEFAULT_ORB_MOTION);
-
-interface RunningStateProps {
+export interface RunningStateProps {
   userName?: string | null;
   userImageUrl?: string | null;
   /** The assistant's generative orb seed — its avatar in every message. */
@@ -143,101 +96,6 @@ interface RunningStateProps {
   /** Re-pull the instance from the orchestrator. SettingsPanel calls this
    * after mutations so the integrations chip / autonomy badge don't drift. */
   onRefresh?: () => void | Promise<void>;
-}
-
-type Message = HermesUiMessage;
-
-const POLL_INTERVAL_MS = 5_000;
-/** Minimum time a reasoning beat stays on screen before the next phase can
- * replace/clear it — so beats don't flash by unreadably. */
-const REASONING_MIN_MS = 1_000;
-
-/**
- * Opt-in flag for Hermes streaming + live progress. Off by default so the chat
- * keeps using the buffered /chat path until the orchestrator side is verified
- * (it must support `stream: true` + emit `event: hermes.status` frames). Flip
- * `NEXT_PUBLIC_HERMES_STREAMING=1` to enable end-to-end streaming.
- */
-const HERMES_STREAMING_ENABLED =
-  process.env.NEXT_PUBLIC_HERMES_STREAMING === "1";
-
-function persistedToMessage(m: HermesPersistedMessage): Message | null {
-  if (m.role !== "user" && m.role !== "assistant") return null;
-  return {
-    id: m.id,
-    role: m.role,
-    content: m.content,
-    kind: m.kind,
-    steps: m.steps,
-    durationMs: m.durationMs,
-    createdAt: m.createdAt,
-  };
-}
-
-function persistedToMessages(messages: HermesPersistedMessage[]): Message[] {
-  return messages
-    .map(persistedToMessage)
-    .filter((m): m is Message => m !== null);
-}
-
-function hasSameMessageIds(left: Message[], right: Message[]): boolean {
-  if (left.length !== right.length) return false;
-  for (let i = 0; i < left.length; i++) {
-    if (left[i]!.id !== right[i]!.id) return false;
-  }
-  return true;
-}
-
-/** Stable-ish key for a turn's duration, surviving the temp→persisted id swap
- * on the post-turn DB re-sync (the content text is unchanged). */
-function durationKey(content: string): string {
-  return content.trim().slice(0, 80);
-}
-
-/** A step in a turn's trace: a `tool` action chip or a `reasoning` beat.
- * Absent `kind` is treated as "tool" (older persisted rows). */
-interface ProgressStep {
-  kind?: "tool" | "reasoning";
-  /** tool_call_id — matches a `tool` frame to its `tool_done`. */
-  id?: string;
-  /** tool: the action label; reasoning: the chain-of-thought snippet. */
-  label: string;
-  detail?: string;
-  /** Set once the tool's `tool_done` frame arrives (chip completes). */
-  done?: boolean;
-}
-
-interface ChatApiResponse {
-  data?: {
-    message?: { role?: string; content?: string };
-    status?: string;
-  };
-  error?: string;
-  message?: string;
-}
-
-function fileToDataUrl(file: File): Promise<string> {
-  return new Promise((resolve, reject) => {
-    const r = new FileReader();
-    r.onload = () => resolve(r.result as string);
-    r.onerror = () => reject(r.error ?? new Error("file_read_failed"));
-    r.readAsDataURL(file);
-  });
-}
-
-/** Prefer the data URL MIME when the browser left `File.type` blank. */
-function mimeFromDataUrl(dataUrl: string): string | null {
-  const match = /^data:([^;,]*);/i.exec(dataUrl);
-  if (!match) return null;
-  const raw = match[1]!.trim();
-  const lower = raw.toLowerCase();
-  if (lower === "" || lower === "application/octet-stream") return null;
-  return lower;
-}
-
-function clientMimeForHermesUpload(file: File, dataUrl: string): string {
-  if (file.type.trim() !== "") return file.type;
-  return mimeFromDataUrl(dataUrl) ?? "application/octet-stream";
 }
 
 export default function RunningState({
@@ -1170,1645 +1028,4 @@ export default function RunningState({
       </AssistantMotionContext.Provider>
     </AssistantSeedContext.Provider>
   );
-}
-
-function WelcomeBlock({ firstName }: { firstName: string | null }) {
-  const t = useTranslations("App.Hermes.Running");
-  const greeting = firstName
-    ? `${t("emptyTitle")}, ${firstName}`
-    : t("emptyTitle");
-
-  // The orchestrator's welcome typically lands within ~2s of arriving here
-  // (it's bundled into the instance "ready" response). Showing the empty
-  // greeting immediately and then replacing it with the real welcome reads
-  // as a glitch — hold the empty state for a moment so it only renders if
-  // there's a true cold start with no welcome incoming.
-  const [show, setShow] = useState(false);
-  useEffect(() => {
-    const t = window.setTimeout(() => setShow(true), 2_500);
-    return () => window.clearTimeout(t);
-  }, []);
-  if (!show) return null;
-
-  return (
-    <div className="mt-[-80px] flex h-full flex-col items-center justify-center px-6">
-      <div className="mx-auto w-full max-w-xl text-center">
-        <h1 className="text-foreground text-3xl font-semibold tracking-tight md:text-4xl">
-          {greeting}
-        </h1>
-        <p className="text-muted-foreground mt-4 text-base leading-relaxed">
-          {t("emptyHint")}
-        </p>
-      </div>
-    </div>
-  );
-}
-
-/**
- * Pull suggested prompts out of a welcome-style message. The orchestrator
- * formats them as markdown list items of the form:
- *   - **Label** — "Quoted prompt to send to Hermes."
- * We extract the quoted text and render it as click-to-send chips below
- * the message. Best-effort: returns [] if no quoted strings are found.
- */
-function extractSuggestedPrompts(content: string): string[] {
-  const prompts: string[] = [];
-  for (const line of content.split("\n")) {
-    if (!line.startsWith("-") && !line.startsWith("*")) continue;
-    // Match the FIRST `"..."` on the line. Use a non-greedy match to handle
-    // multiple quoted segments on one line gracefully (we only chip the first).
-    const match = line.match(/["“]([^"“”]{8,200})["”]/);
-    if (match?.[1]) prompts.push(match[1].trim());
-  }
-  // De-dup while preserving order, cap at 6 to keep the strip readable.
-  return Array.from(new Set(prompts)).slice(0, 6);
-}
-
-function MessageRow({
-  message,
-  userImageUrl,
-  userName,
-  isStreaming = false,
-  durationMs,
-  steps,
-  onSelectSuggestion,
-}: {
-  message: Message;
-  userImageUrl?: string | null;
-  userName?: string | null;
-  isStreaming?: boolean;
-  durationMs?: number;
-  steps?: ProgressStep[];
-  onSelectSuggestion?: (prompt: string) => void;
-}) {
-  const t = useTranslations("App.Hermes.Running");
-  const formatter = useFormatter();
-  const isUser = message.role === "user";
-  const createdAt = new Date(message.createdAt);
-  // Same-day messages show just the time; older ones carry the date so
-  // scrollback stays legible ("Jul 16, 14:32" instead of a bare "14:32").
-  const isSameDay = createdAt.toDateString() === new Date().toDateString();
-  const timestamp = formatter.dateTime(
-    createdAt,
-    isSameDay
-      ? { hour: "2-digit", minute: "2-digit" }
-      : { day: "numeric", month: "short", hour: "2-digit", minute: "2-digit" },
-  );
-
-  if (isUser) {
-    return (
-      <div className="group/message flex w-full justify-end gap-3 px-4 py-0.5">
-        <div className="flex max-w-[70%] flex-col items-end gap-0.5">
-          <div className="bg-muted-foreground/10 text-foreground min-h-6 rounded-lg px-3 py-3 text-sm leading-relaxed whitespace-pre-wrap wrap-break-word">
-            {message.content}
-          </div>
-          <time
-            dateTime={message.createdAt}
-            className="text-tertiary-foreground px-1 text-[10px] tabular-nums opacity-0 transition-opacity group-hover/message:opacity-100"
-          >
-            {timestamp}
-          </time>
-        </div>
-        <Avatar className="size-8 shrink-0">
-          {userImageUrl ? (
-            <AvatarImage
-              src={userImageUrl}
-              alt=""
-              referrerPolicy="no-referrer"
-              onError={(e) => {
-                e.currentTarget.style.display = "none";
-              }}
-            />
-          ) : null}
-          <AvatarFallback className="bg-muted text-muted-foreground text-xs font-medium">
-            {userName?.trim() ? userName.trim().charAt(0).toUpperCase() : "U"}
-          </AvatarFallback>
-        </Avatar>
-      </div>
-    );
-  }
-
-  const chip = describeOutboxKind(message.kind, (key) => t(key));
-  // Only the orchestrator's intro/welcome messages carry suggested prompts.
-  const showSuggestions =
-    onSelectSuggestion !== undefined &&
-    (message.kind === "research_intro" ||
-      message.kind === "welcome" ||
-      message.kind === "returning");
-  const suggestions = showSuggestions
-    ? extractSuggestedPrompts(message.content)
-    : [];
-
-  // Detect orchestrator-pushed "confirmation_resolved" messages with a
-  // sokosumi_create_task payload and split them into prose + task card so
-  // the user doesn't have to read raw JSON in chat.
-  const parsedConfirmation =
-    message.kind === "confirmation_resolved"
-      ? parseConfirmationResolved(message.content, {
-          resolvedFallback: t("confirmation.resolvedFallback"),
-          coworkerFallback: t("confirmation.taskCard.defaultCoworker"),
-          organizationFallback: t("confirmation.taskCard.defaultOrganization"),
-        })
-      : null;
-
-  return (
-    <div className="group/message flex min-h-11 w-full items-start justify-start gap-3 px-4 py-1.5">
-      <AssistantAvatar accent={Boolean(chip)} />
-      <div className="flex min-w-0 flex-1 flex-col gap-1">
-        {chip ? (
-          <span className="border-border/60 text-tertiary-foreground bg-muted/40 inline-flex w-fit items-center rounded-md border px-2 py-0.5 text-xs font-medium uppercase tracking-wider">
-            {chip.label}
-          </span>
-        ) : null}
-        {steps && steps.length > 0 ? (
-          <MessageSteps
-            steps={steps}
-            countLabel={
-              steps.length === 1 && steps[0]!.kind !== "reasoning"
-                ? steps[0]!.label
-                : t("toolSteps", { count: steps.length })
-            }
-          />
-        ) : null}
-        {parsedConfirmation ? (
-          <div className="flex flex-col gap-3 pt-1 pr-10 pb-1">
-            <p className="text-foreground text-sm leading-relaxed">
-              {parsedConfirmation.summary}
-            </p>
-            {parsedConfirmation.task ? (
-              <TaskResultCard task={parsedConfirmation.task} />
-            ) : null}
-          </div>
-        ) : (
-          <Markdown className="text-foreground pt-1 pr-10 pb-1 text-sm">
-            {isStreaming ? `${message.content} ▌` : message.content}
-          </Markdown>
-        )}
-        {suggestions.length > 0 ? (
-          <div className="mt-4 mb-2 flex flex-col gap-2 pr-10 sm:flex-row sm:flex-wrap">
-            {suggestions.map((prompt) => (
-              <button
-                key={prompt}
-                type="button"
-                onClick={() => onSelectSuggestion?.(prompt)}
-                className={cn(
-                  "group/chip border-border bg-card hover:border-foreground/30 hover:bg-muted/40 text-foreground",
-                  "inline-flex max-w-full items-center gap-2.5 rounded-lg border px-4 py-2.5 text-sm font-medium",
-                  "transition-colors active:scale-[0.98]",
-                  "focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2",
-                )}
-              >
-                <span className="truncate text-left">{prompt}</span>
-                <span
-                  aria-hidden
-                  className="text-muted-foreground group-hover/chip:text-primary shrink-0 transition-transform group-hover/chip:translate-x-0.5"
-                >
-                  →
-                </span>
-              </button>
-            ))}
-          </div>
-        ) : null}
-        <div className="flex items-center gap-2 pt-0.5 pb-2">
-          {!isStreaming ? (
-            <CopyButton
-              text={message.content}
-              label={t("copyAction")}
-              copiedLabel={t("copiedAction")}
-            />
-          ) : null}
-          {durationMs !== undefined && !isStreaming ? (
-            <span className="text-tertiary-foreground text-[10px] tabular-nums opacity-0 transition-opacity group-hover/message:opacity-100">
-              {t("answeredIn", { seconds: Math.round(durationMs / 1000) })}
-            </span>
-          ) : null}
-          <time
-            dateTime={message.createdAt}
-            className="text-tertiary-foreground text-[10px] tabular-nums opacity-0 transition-opacity group-hover/message:opacity-100"
-          >
-            {timestamp}
-          </time>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-/** Collapsible disclosure of the tool/progress steps a turn went through —
- * keeps the reasoning legible after the answer has arrived. */
-function MessageSteps({
-  steps,
-  countLabel,
-}: {
-  steps: ProgressStep[];
-  countLabel: string;
-}) {
-  const [open, setOpen] = useState(false);
-  return (
-    <div className="pr-10">
-      <button
-        type="button"
-        onClick={() => setOpen((o) => !o)}
-        aria-expanded={open}
-        className="text-muted-foreground hover:text-foreground focus-visible:ring-primary/40 inline-flex items-center gap-1 rounded text-xs font-medium transition-colors outline-none focus-visible:ring-2"
-      >
-        <Wrench aria-hidden className="size-3" />
-        {countLabel}
-        <ChevronRight
-          aria-hidden
-          className={cn("size-3 transition-transform", open && "rotate-90")}
-        />
-      </button>
-      {open ? (
-        <div className="border-border/60 mt-1.5 flex flex-col gap-1.5 border-l pl-3">
-          {steps.map((step, i) =>
-            step.kind === "reasoning" ? (
-              <p
-                key={`${step.label}-${i}`}
-                className="text-muted-foreground pl-[18px] text-xs italic"
-              >
-                {step.label}
-              </p>
-            ) : (
-              <ToolStepRow key={`${step.label}-${i}`} step={step} />
-            ),
-          )}
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-function ToolStepRow({ step }: { step: ProgressStep }) {
-  return (
-    <div className="text-muted-foreground flex items-start gap-1.5 text-xs">
-      <Check aria-hidden className="text-primary/60 mt-0.5 size-3 shrink-0" />
-      <span className="min-w-0">
-        <span className="text-foreground/80 font-medium">{step.label}</span>
-        {step.detail ? (
-          <span className="text-muted-foreground"> — {step.detail}</span>
-        ) : null}
-      </span>
-    </div>
-  );
-}
-
-/** Hover-to-copy control for an assistant message. */
-function CopyButton({
-  text,
-  label,
-  copiedLabel,
-}: {
-  text: string;
-  label: string;
-  copiedLabel: string;
-}) {
-  const [copied, setCopied] = useState(false);
-  return (
-    <button
-      type="button"
-      aria-label={copied ? copiedLabel : label}
-      onClick={() => {
-        void navigator.clipboard?.writeText(text).then(() => {
-          setCopied(true);
-          setTimeout(() => setCopied(false), 1500);
-        });
-      }}
-      className="text-muted-foreground hover:text-foreground hover:bg-muted/60 border-border/70 focus-visible:ring-primary/40 inline-flex items-center gap-1 rounded-md border px-1.5 py-1 text-[11px] font-medium transition-colors focus-visible:ring-2 focus-visible:outline-none"
-    >
-      {copied ? (
-        <>
-          <Check aria-hidden className="size-3.5" />
-          {copiedLabel}
-        </>
-      ) : (
-        <Copy aria-hidden className="size-3.5" />
-      )}
-    </button>
-  );
-}
-
-interface OutboxKindChip {
-  label: string;
-}
-
-function describeOutboxKind(
-  kind: string | null,
-  t: (key: string) => string,
-): OutboxKindChip | null {
-  if (!kind || kind === "text") return null;
-  if (kind === "welcome" || kind === "research_intro" || kind === "returning") {
-    return { label: t("outboxKinds.welcome") };
-  }
-  if (kind === "daily_brief") return { label: t("outboxKinds.daily_brief") };
-  if (kind === "job_complete") return { label: t("outboxKinds.job_complete") };
-  if (kind === "task_result") return { label: t("outboxKinds.task_result") };
-  if (kind === "daily_suggestions") {
-    return { label: t("outboxKinds.daily_suggestions") };
-  }
-  if (kind === "reminder") return { label: t("outboxKinds.reminder") };
-  if (kind === "confirmation_resolved") {
-    return { label: t("outboxKinds.confirmation_resolved") };
-  }
-  return { label: t("outboxKinds.default") };
-}
-
-/** Transient chain-of-thought beat, shown live and superseded by the next
- * phase frame (held on screen ≥ REASONING_MIN_MS so it doesn't blink). */
-function ReasoningLine({ snippet }: { snippet: string }) {
-  return (
-    <div className="flex w-full items-start gap-3 px-4 pb-1">
-      {/* Gutter aligns under the avatar shown by the indicator above. */}
-      <div className="size-8 shrink-0" aria-hidden />
-      <p className="reasoning-text-shine text-muted-foreground line-clamp-2 max-w-2xl pt-0.5 text-sm italic">
-        {snippet}
-      </p>
-    </div>
-  );
-}
-
-function ProgressChips({
-  chips,
-  startedAt,
-}: {
-  chips: ProgressStep[];
-  startedAt?: number | null;
-}) {
-  return (
-    <div className="flex w-full items-start gap-3 px-4 py-1.5">
-      {/* The active avatar — "focused" eyes while a tool runs, so it reads as
-          working rather than just drafting (the typing indicator stays
-          "thinking"). */}
-      <span className="relative shrink-0">
-        <span
-          aria-hidden
-          className="bg-primary/30 absolute inset-0 animate-ping rounded-full"
-        />
-        <AssistantAvatar animated expression="focused" />
-      </span>
-      <div className="flex min-w-0 flex-col gap-1.5 pt-1">
-        {chips.map((chip, i) => {
-          const isLast = i === chips.length - 1;
-          return (
-            <div
-              key={`${chip.id ?? chip.label}-${i}`}
-              className="flex min-w-0 items-center gap-2 text-sm"
-            >
-              {chip.done ? (
-                <Check
-                  aria-hidden
-                  className="text-primary/70 size-3.5 shrink-0"
-                />
-              ) : (
-                <Loader2
-                  aria-hidden
-                  className="text-primary size-3.5 shrink-0 animate-spin"
-                />
-              )}
-              <span className="text-foreground shrink-0 font-medium">
-                {chip.label}
-              </span>
-              {chip.detail ? (
-                <span className="text-muted-foreground truncate">
-                  {chip.detail}
-                </span>
-              ) : null}
-              {isLast && startedAt ? (
-                <ElapsedTimer startedAt={startedAt} />
-              ) : null}
-            </div>
-          );
-        })}
-      </div>
-    </div>
-  );
-}
-
-/**
- * Pool of "thinking" messages that cycle while Hermes drafts a reply. Mix
- * of straight-faced and lightly silly so users have something to read
- * during long inference runs without it feeling robotic. Each phrase
- * stands on its own — no trailing ellipsis here, the typing dots animate
- * separately. New phrases welcome, just keep them short.
- */
-/** Live "Ns" counter since the turn started — keeps a long wait legible. */
-function ElapsedTimer({ startedAt }: { startedAt: number }) {
-  const [now, setNow] = useState(() => Date.now());
-  useEffect(() => {
-    const id = setInterval(() => setNow(Date.now()), 1000);
-    return () => clearInterval(id);
-  }, []);
-  const secs = Math.max(0, Math.floor((now - startedAt) / 1000));
-  const label =
-    secs < 60
-      ? `${secs}s`
-      : `${Math.floor(secs / 60)}:${String(secs % 60).padStart(2, "0")}`;
-  return (
-    <span className="text-muted-foreground/70 ml-1.5 text-xs tabular-nums">
-      {label}
-    </span>
-  );
-}
-
-function AssistantTyping({ startedAt }: { startedAt?: number | null }) {
-  const t = useTranslations("App.Hermes.Running");
-  const thinkingMessages = orderedMessageList(
-    t.raw("thinkingMessages") as Record<string, string>,
-  );
-  // Escalate reassurance on long waits so a 30s+ turn doesn't read as stuck.
-  const [now, setNow] = useState(() => Date.now());
-  useEffect(() => {
-    if (!startedAt) return;
-    const id = setInterval(() => setNow(Date.now()), 1000);
-    return () => clearInterval(id);
-  }, [startedAt]);
-  const secs = startedAt
-    ? Math.max(0, Math.floor((now - startedAt) / 1000))
-    : 0;
-  const escalation =
-    secs >= 40 ? t("stillWorkingLong") : secs >= 15 ? t("stillWorking") : null;
-
-  return (
-    <div className="flex min-h-11 w-full items-start justify-start gap-3 px-4 py-1.5">
-      {/* Avatar with a slow pulse ring so it reads as "working" at a glance */}
-      <span className="relative shrink-0">
-        <span
-          aria-hidden
-          className="bg-primary/30 absolute inset-0 animate-ping rounded-full"
-        />
-        <AssistantAvatar animated expression="thinking" />
-      </span>
-
-      {/* Rotating phrase + three pulsing dots. Phrase change has its own
-          fade (from RotatingMessages); the dots run independently so there
-          is always something animating even between fades. */}
-      <div className="flex min-h-5 items-center gap-1 pt-2">
-        {escalation ? (
-          <span className="reasoning-text-shine text-foreground text-sm leading-5">
-            {escalation}
-          </span>
-        ) : (
-          <RotatingMessages
-            messages={thinkingMessages}
-            intervalMs={2_800}
-            className="reasoning-text-shine text-foreground text-sm leading-5"
-          />
-        )}
-        <span aria-hidden className="text-foreground/70 inline-flex gap-0.5">
-          <span className="animate-thinking-dot inline-block">.</span>
-          <span className="animate-thinking-dot inline-block [animation-delay:200ms]">
-            .
-          </span>
-          <span className="animate-thinking-dot inline-block [animation-delay:400ms]">
-            .
-          </span>
-        </span>
-        {startedAt ? <ElapsedTimer startedAt={startedAt} /> : null}
-      </div>
-    </div>
-  );
-}
-
-function AssistantAvatar({
-  accent = false,
-  animated = false,
-  expression,
-}: {
-  accent?: boolean;
-  /** Live canvas (one rAF loop) — use only for the single "thinking" avatar. */
-  animated?: boolean;
-  /** Eyes override — "thinking" while it writes. Omit to use the personality's
-   * resting expression. */
-  expression?: OrbExpression;
-} = {}) {
-  const tCommon = useTranslations("App.Hermes.Common");
-  const seed = useContext(AssistantSeedContext);
-  const motion = useContext(AssistantMotionContext);
-  const speed =
-    expression === "thinking" || expression === "focused"
-      ? motion.activeSpeed
-      : motion.speed;
-
-  return (
-    <AuroraOrb
-      seed={seed}
-      size={64}
-      animate={animated}
-      speed={speed}
-      expression={expression ?? motion.restExpression}
-      alt={tCommon("hermesAvatarAlt")}
-      className={cn(
-        "size-8 ring-1",
-        accent ? "ring-border/80" : "ring-border/40",
-      )}
-    />
-  );
-}
-
-interface ComposerProps {
-  ref: React.RefObject<HTMLTextAreaElement | null>;
-  input: string;
-  setInput: (v: string) => void;
-  files: File[];
-  setFiles: (files: File[]) => void;
-  onSubmit: (e: FormEvent<HTMLFormElement>) => void;
-  isReplying: boolean;
-  /** When true (e.g. orchestrator is mid-apply), block the user from sending. */
-  disabled?: boolean;
-  onStop: () => void;
-  placeholder: string;
-  sendLabel: string;
-  stopLabel: string;
-  attachLabel: string;
-}
-
-/**
- * Rotating hints shown in the composer placeholder when the user hasn't
- * typed anything. Gives first-session users concrete things to try without
- * adding chrome below the welcome.
- */
-const ROTATE_INTERVAL_MS = 4_500;
-
-function Composer({
-  ref,
-  input,
-  setInput,
-  files,
-  setFiles,
-  onSubmit,
-  isReplying,
-  disabled = false,
-  onStop,
-  placeholder,
-  sendLabel,
-  stopLabel,
-  attachLabel,
-}: ComposerProps) {
-  const t = useTranslations("App.Hermes.Running");
-  const rotatingHints = orderedMessageList(
-    t.raw("rotatingHints") as Record<string, string>,
-  );
-  const canSend =
-    (input.trim().length > 0 || files.length > 0) && !isReplying && !disabled;
-  const status = isReplying ? "streaming" : "ready";
-
-  // Rotate hint placeholders while the composer is empty + idle. As soon as
-  // the user types or starts a turn, freeze on the default placeholder so we
-  // don't visually fight the typing experience.
-  const [hintIdx, setHintIdx] = useState(0);
-  useEffect(() => {
-    if (input.length > 0 || isReplying) return;
-    const id = window.setInterval(
-      () => setHintIdx((i) => (i + 1) % rotatingHints.length),
-      ROTATE_INTERVAL_MS,
-    );
-    return () => window.clearInterval(id);
-  }, [input.length, isReplying, rotatingHints.length]);
-  const dynamicPlaceholder =
-    input.length > 0 || isReplying
-      ? placeholder
-      : (rotatingHints[hintIdx] ?? placeholder);
-
-  return (
-    <FileUpload
-      value={files}
-      onValueChange={setFiles}
-      multiple
-      maxSize={20 * 1024 * 1024}
-      className="w-full"
-      label={attachLabel}
-    >
-      {/* Same animated gradient border + soft glow as the /chat composer
-          (`.chat-input-border-anchor` in globals.css) so the two chat
-          surfaces read as one product. */}
-      <div
-        className={cn(
-          "chat-input-border-anchor",
-          // z-10 lifts the glow above the scroll-fade overlay (z-5) — without
-          // it the fade's opaque bottom edge masks the halo into a hard line.
-          "relative z-10 rounded-xl",
-          "shadow-[0_0_16px_0] shadow-primary/15",
-          "focus-within:shadow-[0_0_24px_2px] focus-within:shadow-primary/30",
-          "transition-shadow duration-300",
-          disabled && "pointer-events-none opacity-60",
-        )}
-      >
-        <PromptInput
-          onSubmit={onSubmit}
-          className="bg-background relative z-10 rounded-[calc(var(--radius-xl)-1.5px)] border-0 shadow-none transition-all duration-200"
-        >
-          <FileUploadDropzone
-            className="data-dragging:bg-accent/20 w-full items-stretch justify-start border-0 p-0 hover:bg-transparent"
-            onClick={(event) => event.preventDefault()}
-          >
-            <PromptInputTextarea
-              ref={ref}
-              value={input}
-              onChange={(e) => setInput(e.target.value)}
-              placeholder={
-                disabled ? t("composerDisabledPlaceholder") : dynamicPlaceholder
-              }
-              disableAutoResize
-              maxHeight={200}
-              minHeight={44}
-              autoFocus
-              disabled={disabled}
-              className="placeholder:text-muted-foreground scrollbar-none grow resize-none border-0! bg-transparent p-4 text-base ring-0 outline-none [-ms-overflow-style:none] focus-visible:ring-0 focus-visible:ring-offset-0 focus-visible:outline-none [&::-webkit-scrollbar]:hidden"
-            />
-          </FileUploadDropzone>
-
-          <FileUploadList orientation="horizontal" className="gap-2 px-3 pb-1">
-            {files.map((file) => (
-              <FileUploadItem
-                key={`${file.name}-${file.lastModified}`}
-                value={file}
-                className="bg-muted/40 border-border/60 flex max-w-56 items-center gap-2 rounded-md border px-2 py-1.5"
-              >
-                <FileUploadItemPreview className="size-7 shrink-0 rounded" />
-                <FileUploadItemMetadata className="min-w-0 flex-1 text-xs" />
-                <FileUploadItemDelete asChild>
-                  <Button
-                    type="button"
-                    variant="ghost"
-                    size="icon"
-                    className="text-muted-foreground hover:text-foreground size-6 shrink-0 rounded-full"
-                    aria-label={t("removeFile")}
-                  >
-                    <X className="size-3.5" aria-hidden />
-                  </Button>
-                </FileUploadItemDelete>
-              </FileUploadItem>
-            ))}
-          </FileUploadList>
-
-          <PromptInputToolbar className="border-t-0 p-3">
-            <PromptInputTools className="flex-wrap gap-1 sm:gap-1.5">
-              <FileUploadTrigger asChild>
-                <Button
-                  type="button"
-                  variant="ghost"
-                  size="sm"
-                  className="size-8 rounded-full! p-0"
-                  title={attachLabel}
-                  aria-label={attachLabel}
-                >
-                  <Plus className="size-3.5" />
-                </Button>
-              </FileUploadTrigger>
-            </PromptInputTools>
-
-            {isReplying ? (
-              <Button
-                type="button"
-                variant="default"
-                size="icon"
-                onClick={onStop}
-                aria-label={stopLabel}
-                className="size-8 rounded-full"
-              >
-                <StopIcon size={14} />
-              </Button>
-            ) : (
-              <PromptInputSubmit
-                className="size-8 rounded-full transition-colors duration-200"
-                disabled={!canSend}
-                status={status}
-                aria-label={sendLabel}
-              >
-                <ArrowUpIcon size={14} />
-              </PromptInputSubmit>
-            )}
-          </PromptInputToolbar>
-        </PromptInput>
-      </div>
-    </FileUpload>
-  );
-}
-
-const INTEGRATION_ICON_BY_PROVIDER: Record<HermesIntegrationProvider, string> =
-  {
-    gmail: "/icons/gmail.svg",
-    google_calendar: "/icons/google-calendar.svg",
-    google_sheets: "/icons/google-sheets.svg",
-    google_docs: "/icons/google-docs.svg",
-    outlook: "/icons/outlook.svg",
-    outlook_calendar: "/icons/outlook.svg",
-    slack: "/icons/slack.svg",
-    teams: "/icons/teams.svg",
-    linear: "/icons/linear.svg",
-    jira: "/icons/jira.svg",
-    github: "/icons/github.svg",
-    notion: "/icons/notion.svg",
-    hubspot: "/icons/hubspot.svg",
-    twitter: "/icons/x.svg",
-    instagram: "/icons/instagram.svg",
-    youtube: "/icons/youtube.svg",
-    linkedin: "/icons/linkedin.svg",
-  };
-
-/**
- * Some Composio toolkits cover multiple orchestrator provider strings
- * from a single OAuth (Outlook's mail + calendar share one consent and
- * land as two integration rows). The chat chip should treat those as
- * one connected service so the count + icon stack reflect reality.
- */
-function canonicalServiceKey(provider: HermesIntegrationProvider): string {
-  if (provider === "outlook" || provider === "outlook_calendar") {
-    return "outlook";
-  }
-  return provider;
-}
-
-function dedupeServiceIntegrations(
-  integrations: HermesIntegrationPublic[],
-): HermesIntegrationPublic[] {
-  const seen = new Set<string>();
-  const result: HermesIntegrationPublic[] = [];
-  for (const integration of integrations) {
-    const key = canonicalServiceKey(integration.provider);
-    if (seen.has(key)) continue;
-    seen.add(key);
-    result.push(integration);
-  }
-  return result;
-}
-
-/**
- * Top-right chip opening the Autonomy panel — the home for "what the
- * assistant does on its own" (autonomy level + scheduled tasks). Styled to
- * match IntegrationsChip so the two controls read as one cluster.
- */
-function AutonomyChip({ onClick }: { onClick: () => void }) {
-  const tPanel = useTranslations("App.Hermes.AutonomyPanel");
-
-  return (
-    <Tooltip>
-      <TooltipTrigger asChild>
-        <button
-          type="button"
-          onClick={onClick}
-          className="border-border bg-card text-foreground hover:bg-muted/40 hover:border-foreground/30 inline-flex h-8 items-center gap-2 rounded-full border px-2.5 text-xs font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
-          aria-label={tPanel("title")}
-        >
-          <Zap className="text-tertiary-foreground size-3.5" aria-hidden />
-          <span>{tPanel("chip")}</span>
-        </button>
-      </TooltipTrigger>
-      <TooltipContent side="bottom">{tPanel("subtitle")}</TooltipContent>
-    </Tooltip>
-  );
-}
-
-function IntegrationsChip({
-  integrations,
-  onClick,
-}: {
-  integrations: HermesIntegrationPublic[];
-  onClick: () => void;
-}) {
-  const t = useTranslations("App.Hermes.Running.integrationsChip");
-  const tRunning = useTranslations("App.Hermes.Running");
-  // Dedupe paired providers (outlook + outlook_calendar share one OAuth)
-  // so the chip shows one entry per real service. Otherwise a single
-  // Outlook connection looks like "2 connected" with the same icon twice.
-  const connected = dedupeServiceIntegrations(
-    integrations.filter((i) => i.status === "connected"),
-  );
-  const stacked = connected.slice(0, 3);
-  const hasAny = connected.length > 0;
-
-  return (
-    <Tooltip>
-      <TooltipTrigger asChild>
-        <button
-          type="button"
-          onClick={onClick}
-          className="border-border bg-card text-foreground hover:bg-muted/40 hover:border-foreground/30 inline-flex h-8 items-center gap-2 rounded-full border pl-1.5 pr-2.5 text-xs font-medium transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2"
-          aria-label={
-            hasAny
-              ? t("ariaConnected", { count: connected.length })
-              : tRunning("settingsCta")
-          }
-        >
-          {hasAny ? (
-            <>
-              <span className="flex items-center -space-x-1.5">
-                {stacked.map((i) => (
-                  <span
-                    key={i.provider}
-                    className="border-card bg-background inline-flex size-5 items-center justify-center rounded-full border-2"
-                  >
-                    {/* eslint-disable-next-line @next/next/no-img-element */}
-                    <img
-                      src={INTEGRATION_ICON_BY_PROVIDER[i.provider]}
-                      alt=""
-                      className="size-3"
-                    />
-                  </span>
-                ))}
-              </span>
-              <span className="tabular-nums">{connected.length}</span>
-            </>
-          ) : (
-            <>
-              <SettingsIcon
-                className="text-tertiary-foreground size-3.5"
-                aria-hidden
-              />
-              <span>{tRunning("settingsCta")}</span>
-            </>
-          )}
-        </button>
-      </TooltipTrigger>
-      <TooltipContent side="bottom">
-        {hasAny ? t("integrations") : tRunning("settingsCta")}
-      </TooltipContent>
-    </Tooltip>
-  );
-}
-
-type HermesIntegrationPublic = NonNullable<
-  RunningStateProps["instance"]
->["integrations"][number];
-type HermesIntegrationProvider = HermesIntegrationPublic["provider"];
-
-/**
- * Same RFC-4122 UUID pattern the server uses to resolve coworker /
- * organization ids in the summary. Splitting on this lets us interleave
- * `<CoworkerRefChip>` / `<OrgRefChip>` exactly where the orchestrator
- * wrote the id.
- */
-const SUMMARY_UUID_PATTERN =
-  /([0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12})/gi;
-
-/**
- * Dev preview: synthesizes a fake pending confirmation referencing a real
- * coworker so the ConfirmationCard renders with avatar + name chips
- * locally, without waiting on the orchestrator. Activated by
- * `?state=running&mock=confirmation`. Optional overrides via
- * `&toolName=sokosumi_create_job&assigneeId=<uuid>&coworkerName=<name>&coworkerImage=<url>`.
- * Legacy `coworkerId` is still accepted for bookmarked demo URLs.
- */
-function buildMockPendingConfirmations(
-  params: Pick<URLSearchParams, "get">,
-): HermesPendingConfirmation[] {
-  if (params.get("mock") !== "confirmation") return [];
-  const requestedToolName = params.get("toolName");
-  const toolName =
-    requestedToolName && isConfirmationOrgAwareTool(requestedToolName)
-      ? requestedToolName
-      : "sokosumi_create_task";
-  const assigneeId =
-    params.get("assigneeId") ??
-    params.get("coworkerId") ??
-    "0e8c93b0-5332-4734-b603-ea18d17b50c5";
-  const coworkerName = params.get("coworkerName") ?? "Hannah";
-  const coworkerImage = params.get("coworkerImage");
-  const summary =
-    toolName === "sokosumi_create_job"
-      ? `Create a new job "Research: Teodor Petricevic — UNDP AltFinLab" and assign it to coworker ${assigneeId}.`
-      : `Create a new task "Research: Teodor Petricevic — UNDP AltFinLab" and assign it to coworker ${assigneeId}.`;
-  return [
-    {
-      id: "mock-confirmation-1",
-      toolName,
-      summary,
-      createdAt: new Date().toISOString(),
-      referencedCoworkers: [
-        {
-          id: assigneeId,
-          name: coworkerName,
-          image: coworkerImage,
-        },
-      ],
-      referencedOrganizations: [],
-      // Optional preview overrides for the proposed-workspace dropdown default.
-      organizationId: params.get("organizationId"),
-      organizationName: params.get("organizationName"),
-    },
-  ];
-}
-
-function renderConfirmationSummary(
-  confirmation: HermesPendingConfirmation,
-): React.ReactNode {
-  const { summary, referencedCoworkers, referencedOrganizations } =
-    confirmation;
-  if (
-    referencedCoworkers.length === 0 &&
-    referencedOrganizations.length === 0
-  ) {
-    return summary;
-  }
-  const coworkerById = new Map(
-    referencedCoworkers.map((c) => [c.id.toLowerCase(), c]),
-  );
-  const orgById = new Map(
-    referencedOrganizations.map((o) => [o.id.toLowerCase(), o]),
-  );
-
-  return summary.split(SUMMARY_UUID_PATTERN).map((part, index) => {
-    const key = `${index}-${part}`;
-    const lower = part.toLowerCase();
-    const coworker = coworkerById.get(lower);
-    if (coworker) {
-      return <CoworkerRefChip key={key} coworker={coworker} />;
-    }
-    const organization = orgById.get(lower);
-    if (organization) {
-      return <OrgRefChip key={key} organization={organization} />;
-    }
-    // Unresolved chunk — either non-UUID prose or an id we couldn't
-    // attribute to this user. Render verbatim.
-    return <React.Fragment key={key}>{part}</React.Fragment>;
-  });
-}
-
-function CoworkerRefChip({
-  coworker,
-}: {
-  coworker: HermesPendingConfirmation["referencedCoworkers"][number];
-}) {
-  return (
-    <span className="border-border/60 bg-card/80 text-foreground mx-0.5 inline-flex max-w-56 items-center gap-1.5 rounded-md border px-1.5 py-0.5 align-middle text-xs font-medium">
-      {coworker.image ? (
-        /* eslint-disable-next-line @next/next/no-img-element */
-        <img
-          src={coworker.image}
-          alt=""
-          className="border-border size-4 shrink-0 rounded-full border"
-        />
-      ) : (
-        <span className="bg-muted text-muted-foreground inline-flex size-4 shrink-0 items-center justify-center rounded-full text-[10px] font-semibold">
-          {coworker.name.charAt(0).toUpperCase()}
-        </span>
-      )}
-      <span className="truncate">{coworker.name}</span>
-    </span>
-  );
-}
-
-function OrgRefChip({
-  organization,
-}: {
-  organization: HermesPendingConfirmation["referencedOrganizations"][number];
-}) {
-  return (
-    <span className="border-border/60 bg-card/80 text-foreground mx-0.5 inline-flex max-w-56 items-center gap-1.5 rounded-md border px-1.5 py-0.5 align-middle text-xs font-medium">
-      <Building2
-        className="text-muted-foreground size-3.5 shrink-0"
-        aria-hidden
-      />
-      <span className="truncate">{organization.name}</span>
-    </span>
-  );
-}
-
-/**
- * Inline approve/reject card for medium-autonomy gates. Approve/reject only
- * move the card into the read-only audit trail when the orchestrator reports
- * the matching terminal status (`approved` / `rejected`). When the
- * orchestrator returns `status === "errored"` (HTTP 200), we show a toast and
- * leave the card interactive so the user can retry. `already_resolved` and
- * the opposite resolution on either action still settle the card — the gate
- * was handled elsewhere (another tab, stale list, etc.).
- */
-/**
- * Tools that may spend credits. `sokosumi_create_job` always does; tasks
- * spend once a coworker actually runs against them. Either way we want
- * the user to see the "costs deduct from credits" notice before they
- * approve.
- */
-const COST_BEARING_TOOLS = new Set([
-  "sokosumi_create_task",
-  "sokosumi_create_job",
-]);
-
-/**
- * Captures what the user did with a confirmation so the card can be
- * re-rendered as a read-only audit trail. `organizationId === undefined`
- * means the tool wasn't org-aware (no dropdown was shown); `null` means
- * the user explicitly picked personal scope.
- */
-interface ConfirmationResolution {
-  status: "approved" | "rejected" | "already_resolved";
-  organizationId?: string | null;
-}
-
-type TimelineEntry =
-  | { kind: "message"; ts: number; key: string; message: Message }
-  | {
-      kind: "resolved";
-      ts: number;
-      key: string;
-      entry: ResolvedConfirmationEntry;
-    };
-
-interface ResolvedConfirmationEntry {
-  confirmation: HermesPendingConfirmation;
-  resolution: ConfirmationResolution;
-  /**
-   * Fixed timeline position: just past the newest message visible at the
-   * moment the user resolved the card. Snapshotted once so the card stays
-   * exactly where the user acted — mixing the gate's server `createdAt`
-   * with client-clock message timestamps let the card teleport above the
-   * user's own (optimistically-stamped) message on approval, then strand
-   * far up in the scrollback once the post-turn re-sync rewrote message
-   * ids/timestamps to server time.
-   */
-  timelineTs: number;
-}
-
-function ConfirmationCard({
-  confirmation,
-  onResolved,
-  organizations,
-  activeOrganizationId,
-  resolution,
-  hasActiveSubscription = true,
-  onRequireSubscription,
-}: {
-  confirmation: HermesPendingConfirmation;
-  onResolved: (
-    confirmationId: string,
-    resolution: ConfirmationResolution,
-    confirmation: HermesPendingConfirmation,
-  ) => void;
-  organizations: HermesOrganizationOption[];
-  activeOrganizationId: string | null;
-  /** Non-null means the user already resolved this card; render read-only. */
-  resolution: ConfirmationResolution | null;
-  hasActiveSubscription?: boolean;
-  onRequireSubscription?: () => void;
-}) {
-  const t = useTranslations("App.Hermes.Running.confirmation");
-  const [busy, setBusy] = useState<"approving" | "rejecting" | null>(null);
-
-  const isResolved = resolution !== null;
-  const showOrgPicker = isConfirmationOrgAwareTool(confirmation.toolName);
-  const showCostNotice = COST_BEARING_TOOLS.has(confirmation.toolName);
-  const orgPickerOptions = mergeConfirmationOrgPickerOptions(
-    organizations,
-    confirmation,
-  );
-  const initialOrgValue =
-    resolution && resolution.organizationId !== undefined
-      ? (resolution.organizationId ?? CONFIRMATION_PERSONAL_SCOPE_VALUE)
-      : resolveConfirmationOrgPickerValue(
-          confirmation,
-          organizations,
-          activeOrganizationId,
-        );
-  const proposedOrgValue = resolveConfirmationOrgPickerValue(
-    confirmation,
-    organizations,
-    activeOrganizationId,
-  );
-  const [orgSelection, setOrgSelection] = useState(() => ({
-    baselineOrgValue: initialOrgValue,
-    selectedOrgValue: initialOrgValue,
-    userChangedOrg: false,
-  }));
-  const selectedOrgRef = useRef(orgSelection.selectedOrgValue);
-  selectedOrgRef.current = orgSelection.selectedOrgValue;
-
-  if (!isResolved && showOrgPicker) {
-    const syncedOrgSelection = applyConfirmationOrgProposalUpdate(
-      proposedOrgValue,
-      orgSelection,
-    );
-    if (syncedOrgSelection !== orgSelection) {
-      setOrgSelection(syncedOrgSelection);
-      selectedOrgRef.current = syncedOrgSelection.selectedOrgValue;
-    }
-  }
-
-  const selectedOrgValue = orgSelection.selectedOrgValue;
-
-  const handleOrgValueChange = (value: string) => {
-    selectedOrgRef.current = value;
-    setOrgSelection((current) => ({
-      ...current,
-      selectedOrgValue: value,
-      userChangedOrg: true,
-    }));
-  };
-
-  const handleApprove = async () => {
-    if (busy || isResolved) return;
-    if (!hasActiveSubscription) {
-      onRequireSubscription?.();
-      return;
-    }
-    setBusy("approving");
-    // The workspace dropdown shows a local default that may NOT match the
-    // workspace Hermes proposed in its tool call. Only send an organization
-    // override when the user actually changes the dropdown; if they leave it
-    // untouched, omit the field entirely so Hermes' proposed workspace stands.
-    // Sending `organizationId` (incl. `null` for Personal) on an untouched
-    // dropdown asserts a workspace choice the user never made and clobbers
-    // Hermes' selection — e.g. filing a task in Personal instead of the org
-    // Hermes chose. The dropdown default is reconciled with Hermes' actual
-    // proposal separately (pending orchestrator-provided proposed workspace).
-    // Current dropdown selection — used for the resolved-card audit display.
-    const workspaceSelection = showOrgPicker
-      ? buildCurrentConfirmationApproveOrganizationOverride(
-          selectedOrgRef,
-          orgPickerOptions,
-        )
-      : undefined;
-    // Override actually sent: only when the user changed the workspace, so an
-    // untouched dropdown leaves Hermes' proposed workspace intact.
-    const workspaceOverride = showOrgPicker
-      ? buildConfirmationApproveOverrideIfChanged(
-          selectedOrgRef.current,
-          orgSelection.baselineOrgValue,
-          orgPickerOptions,
-        )
-      : undefined;
-    const result = await approveHermesConfirmationAction(
-      workspaceOverride
-        ? {
-            confirmationId: confirmation.id,
-            confirmation,
-            ...workspaceOverride,
-          }
-        : { confirmationId: confirmation.id, confirmation },
-    );
-    setBusy(null);
-    if (!result.ok) {
-      toast.error(result.error.message ?? t("approveFailed"));
-      return;
-    }
-    const { status } = result.data;
-    const resolutionOrgId = workspaceSelection?.organizationId;
-
-    if (status === "errored") {
-      toast.error(result.data.error ?? t("erroredAfterApproval"));
-      return;
-    }
-    if (status === "approved") {
-      toast.success(t("approvedToast"));
-      onResolved(
-        confirmation.id,
-        {
-          status: "approved",
-          organizationId: resolutionOrgId,
-        },
-        confirmation,
-      );
-      return;
-    }
-    if (status === "already_resolved") {
-      toast.info(t("alreadyResolvedToast"));
-      onResolved(
-        confirmation.id,
-        {
-          status: "already_resolved",
-          organizationId: resolutionOrgId,
-        },
-        confirmation,
-      );
-      return;
-    }
-    if (status === "rejected") {
-      toast.info(t("alreadyResolvedToast"));
-      onResolved(
-        confirmation.id,
-        {
-          status: "rejected",
-          organizationId: resolutionOrgId,
-        },
-        confirmation,
-      );
-    }
-  };
-
-  const handleReject = async () => {
-    if (busy || isResolved) return;
-    if (!hasActiveSubscription) {
-      onRequireSubscription?.();
-      return;
-    }
-    setBusy("rejecting");
-    const resolutionOrgId = showOrgPicker
-      ? buildCurrentConfirmationApproveOrganizationOverride(
-          selectedOrgRef,
-          orgPickerOptions,
-        ).organizationId
-      : undefined;
-    const result = await rejectHermesConfirmationAction({
-      confirmationId: confirmation.id,
-      confirmation,
-    });
-    setBusy(null);
-    if (!result.ok) {
-      toast.error(result.error.message ?? t("rejectFailed"));
-      return;
-    }
-    const { status } = result.data;
-
-    if (status === "errored") {
-      toast.error(result.data.error ?? t("rejectFailed"));
-      return;
-    }
-    if (status === "rejected") {
-      toast.success(t("rejectedToast"));
-      onResolved(
-        confirmation.id,
-        {
-          status: "rejected",
-          organizationId: resolutionOrgId,
-        },
-        confirmation,
-      );
-      return;
-    }
-    if (status === "already_resolved") {
-      toast.info(t("alreadyResolvedToast"));
-      onResolved(
-        confirmation.id,
-        {
-          status: "already_resolved",
-          organizationId: resolutionOrgId,
-        },
-        confirmation,
-      );
-      return;
-    }
-    if (status === "approved") {
-      toast.info(t("alreadyResolvedToast"));
-      onResolved(
-        confirmation.id,
-        {
-          status: "approved",
-          organizationId: resolutionOrgId,
-        },
-        confirmation,
-      );
-    }
-  };
-
-  const tool = describeConfirmationTool(confirmation.toolName, (key) => t(key));
-  const summaryFragments = renderConfirmationSummary(confirmation);
-
-  // Resolved cards: same layout, muted chrome, status pill instead of
-  // buttons, dropdown locked to the user's earlier choice. Renders as a
-  // read-only audit trail in the chat.
-  const isApproved = resolution?.status === "approved";
-  const isAlreadyResolved = resolution?.status === "already_resolved";
-
-  return (
-    <div className="flex min-h-11 w-full items-start justify-start gap-3 px-4 py-1.5">
-      <AssistantAvatar accent={!isResolved} />
-      <div
-        className={cn(
-          "flex min-w-0 flex-1 flex-col gap-3 rounded-2xl border px-4 py-3 backdrop-blur-sm",
-          isResolved
-            ? "border-border/60 bg-muted/30"
-            : "border-amber-500/30 bg-amber-500/6",
-        )}
-      >
-        <div
-          className={cn(
-            "inline-flex items-center gap-2 text-xs font-semibold uppercase tracking-wider",
-            isResolved
-              ? "text-muted-foreground"
-              : "text-amber-700 dark:text-amber-400",
-          )}
-        >
-          {isResolved ? (
-            isApproved || isAlreadyResolved ? (
-              <Check className="size-3.5" aria-hidden />
-            ) : (
-              <X className="size-3.5" aria-hidden />
-            )
-          ) : (
-            <AlertCircle className="size-3.5" aria-hidden />
-          )}
-          <span>
-            {isResolved
-              ? isApproved
-                ? t("approvedStatus")
-                : isAlreadyResolved
-                  ? t("alreadyResolvedStatus")
-                  : t("rejectedStatus")
-              : t("heading")}
-          </span>
-        </div>
-        <div className="flex flex-col gap-1">
-          <div
-            className={cn(
-              "text-sm font-semibold tracking-tight",
-              isResolved ? "text-muted-foreground" : "text-foreground",
-            )}
-          >
-            {tool.action}
-          </div>
-          <p
-            className={cn(
-              "text-sm leading-relaxed",
-              isResolved ? "text-muted-foreground" : "text-foreground/90",
-            )}
-          >
-            {summaryFragments}
-          </p>
-        </div>
-        {isResolved &&
-        isApproved &&
-        confirmation.toolName === "sokosumi_create_task" ? (
-          <div className="flex items-center gap-2 text-xs text-emerald-600 dark:text-emerald-400">
-            <Check className="size-3.5 shrink-0" aria-hidden />
-            <span>{t("creatingInBackground")}</span>
-          </div>
-        ) : null}
-        {showOrgPicker ? (
-          <div className="flex flex-col gap-1.5">
-            <label
-              htmlFor={`confirm-org-${confirmation.id}`}
-              className="text-muted-foreground text-xs font-medium uppercase tracking-wider"
-            >
-              {t("organizationLabel")}
-            </label>
-            <Select
-              value={selectedOrgValue}
-              onValueChange={handleOrgValueChange}
-              disabled={busy !== null || isResolved}
-            >
-              <SelectTrigger
-                id={`confirm-org-${confirmation.id}`}
-                size="sm"
-                className="w-full max-w-xs"
-              >
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value={CONFIRMATION_PERSONAL_SCOPE_VALUE}>
-                  {t("organizationPersonal")}
-                </SelectItem>
-                {orgPickerOptions.map((org) => (
-                  <SelectItem key={org.id} value={org.id}>
-                    {org.name}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-          </div>
-        ) : null}
-        {showCostNotice && !isResolved ? (
-          <div className="border-border/60 bg-muted/30 text-muted-foreground flex items-start gap-2 rounded-md border px-2.5 py-2 text-xs leading-relaxed">
-            <Coins className="mt-0.5 size-3.5 shrink-0" aria-hidden />
-            <span>{t("costNotice")}</span>
-          </div>
-        ) : null}
-        {isResolved ? null : (
-          <div className="flex items-center gap-2 pt-1">
-            <Button
-              type="button"
-              size="sm"
-              variant="primary"
-              className="gap-1.5"
-              disabled={busy !== null}
-              onClick={() => void handleApprove()}
-            >
-              {busy === "approving" ? (
-                <Loader2 className="size-3.5 animate-spin" aria-hidden />
-              ) : (
-                <Check className="size-3.5" aria-hidden />
-              )}
-              <span>{t("approve")}</span>
-            </Button>
-            <Button
-              type="button"
-              size="sm"
-              variant="outline"
-              className="gap-1.5"
-              disabled={busy !== null}
-              onClick={() => void handleReject()}
-            >
-              {busy === "rejecting" ? (
-                <Loader2 className="size-3.5 animate-spin" aria-hidden />
-              ) : (
-                <X className="size-3.5" aria-hidden />
-              )}
-              <span>{t("reject")}</span>
-            </Button>
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
-
-// ─────────────────────────────────────────────────────────────────────────────
-// Confirmation-resolved message rendering
-//
-// The orchestrator pushes a `confirmation_resolved` outbox message after the
-// user approves a tool call, with body shaped like:
-//
-//   The user approved your earlier sokosumi_create_task request. The action
-//   was executed; here's the result you can act on:
-//   { ...big JSON blob... }
-//
-// Dumping that JSON in chat is hostile. Parse it, render the prose intro,
-// and if the payload is a known shape (currently sokosumi_create_task)
-// render a Task Card linking to /tasks/:id instead.
-
-interface ParsedTaskResult {
-  id: string;
-  name: string;
-  description: string | null;
-  status: string | null;
-  credits: number | null;
-  coworker: { name: string; image: string | null } | null;
-  organization: { name: string; slug: string | null } | null;
-}
-
-interface ParsedConfirmationResolved {
-  /** Prose lead-in with the JSON block stripped out. */
-  summary: string;
-  /** Populated when the JSON payload matched the sokosumi_create_task shape. */
-  task: ParsedTaskResult | null;
-}
-
-function parseConfirmationResolved(
-  content: string,
-  fallbacks: {
-    resolvedFallback: string;
-    coworkerFallback: string;
-    organizationFallback: string;
-  },
-): ParsedConfirmationResolved | null {
-  if (!content) return null;
-  // Find the first opening brace at the start of a line, take everything
-  // from there as the JSON region. The intro prose is whatever comes before.
-  const braceIdx = content.search(/^\s*{/m);
-  if (braceIdx < 0) {
-    // No JSON in the body — just return the message as-is so the caller
-    // can render it as plain markdown again.
-    return null;
-  }
-  const summary = content.slice(0, braceIdx).trim();
-  const rawJson = content.slice(braceIdx).trim();
-
-  let parsed: unknown;
-  try {
-    parsed = JSON.parse(rawJson);
-  } catch {
-    return { summary: summary || content, task: null };
-  }
-
-  return {
-    summary: summary || fallbacks.resolvedFallback,
-    task: extractTaskFromConfirmation(parsed, fallbacks),
-  };
-}
-
-function extractTaskFromConfirmation(
-  payload: unknown,
-  fallbacks: {
-    coworkerFallback: string;
-    organizationFallback: string;
-  },
-): ParsedTaskResult | null {
-  if (!payload || typeof payload !== "object") return null;
-  const root = payload as Record<string, unknown>;
-  const taskWrapper = root.task as Record<string, unknown> | undefined;
-  const data = (taskWrapper?.data ?? taskWrapper) as
-    | Record<string, unknown>
-    | undefined;
-  if (!data || typeof data !== "object") return null;
-
-  const id = typeof data.id === "string" ? data.id : null;
-  const name = typeof data.name === "string" ? data.name : null;
-  if (!id || !name) return null;
-
-  const coworker = (data.assignee ?? data.coworker) as
-    | Record<string, unknown>
-    | null
-    | undefined;
-  const organization = data.organization as
-    | Record<string, unknown>
-    | null
-    | undefined;
-
-  return {
-    id,
-    name,
-    description: typeof data.description === "string" ? data.description : null,
-    status: typeof data.status === "string" ? data.status : null,
-    credits: typeof data.credits === "number" ? data.credits : null,
-    coworker: coworker
-      ? {
-          name:
-            typeof coworker.name === "string"
-              ? coworker.name
-              : fallbacks.coworkerFallback,
-          image: typeof coworker.image === "string" ? coworker.image : null,
-        }
-      : null,
-    organization: organization
-      ? {
-          name:
-            typeof organization.name === "string"
-              ? organization.name
-              : fallbacks.organizationFallback,
-          slug:
-            typeof organization.slug === "string" ? organization.slug : null,
-        }
-      : null,
-  };
-}
-
-/**
- * A compact card for `sokosumi_create_task` results pushed via
- * `confirmation_resolved`. Replaces what would otherwise be a 60-line raw
- * JSON dump with the bits a human actually wants: name, who it's assigned
- * to, status, and a deep link.
- */
-function TaskResultCard({ task }: { task: ParsedTaskResult }) {
-  const t = useTranslations("App.Hermes.Running.confirmation.taskCard");
-
-  return (
-    <Link
-      href={`/tasks/${task.id}`}
-      className="border-border bg-card/60 hover:border-foreground/30 hover:bg-card group/task-card flex max-w-2xl flex-col gap-3 rounded-2xl border p-4 transition-colors"
-    >
-      <div className="flex items-center gap-2">
-        {task.coworker?.image ? (
-          /* eslint-disable-next-line @next/next/no-img-element */
-          <img
-            src={task.coworker.image}
-            alt=""
-            className="border-border size-6 shrink-0 rounded-full border"
-          />
-        ) : task.coworker ? (
-          <span className="bg-muted text-muted-foreground inline-flex size-6 shrink-0 items-center justify-center rounded-full text-xs font-medium">
-            {task.coworker.name.charAt(0).toUpperCase()}
-          </span>
-        ) : null}
-        <span className="text-foreground text-sm font-medium">
-          {task.coworker?.name ?? t("defaultTask")}
-        </span>
-        {task.organization ? (
-          <>
-            <span className="text-tertiary-foreground text-xs">·</span>
-            <span className="text-muted-foreground text-xs">
-              {task.organization.name}
-            </span>
-          </>
-        ) : null}
-        {task.status ? (
-          <span className="border-border/60 text-muted-foreground ml-auto rounded-full border px-1.5 py-0.5 text-xs font-medium uppercase tracking-wider">
-            {task.status.toLowerCase()}
-          </span>
-        ) : null}
-      </div>
-
-      <div>
-        <div className="text-foreground text-base font-semibold tracking-tight">
-          {task.name}
-        </div>
-        {task.description ? (
-          <p className="text-muted-foreground mt-1 line-clamp-3 text-sm leading-relaxed">
-            {task.description}
-          </p>
-        ) : null}
-      </div>
-
-      <div className="text-muted-foreground group-hover/task-card:text-foreground inline-flex items-center gap-1 text-xs font-medium transition-colors">
-        <span>{t("viewTask")}</span>
-        <ArrowUpRight className="size-3.5" aria-hidden />
-      </div>
-    </Link>
-  );
-}
-
-/**
- * Maps a tool slug to user-facing copy for the confirmation card.
- * Hides the technical sokosumi_* prefix; falls back to the raw slug for
- * future kinds.
- */
-const CONFIRMATION_TOOL_KEYS = [
-  "sokosumi_create_task",
-  "sokosumi_create_job",
-  "sokosumi_add_task_comment",
-  "sokosumi_provide_job_input",
-  "sokosumi_refund_job",
-] as const;
-
-function describeConfirmationTool(
-  toolName: string,
-  t: (key: string) => string,
-): {
-  action: string;
-  helper: string;
-} {
-  if ((CONFIRMATION_TOOL_KEYS as readonly string[]).includes(toolName)) {
-    return {
-      action: t(`tools.${toolName}.action`),
-      helper: t(`tools.${toolName}.helper`),
-    };
-  }
-  return { action: toolName, helper: "" };
 }
